rewrite to reflect correct file structure
rewriting files and got context menu working... sorta. the context option for SeeColor is on the right click menu. Currently the option does nothing, trying to figure out why it doesn't load the image elements. the console log option is there to throw out output when it does image sourcing.