Finalized version of the Firefox extension
As stated in the readme file, this extension does not work with newer versions of the firefox browser and should not be used due to the security bypassing required to allow the scripts to run. Manifest v3 has many changes to the firefox content security policy and as such requires that extensions do not use cross site scripting methods as we use in the chrome version. The extension performs the same as the chrome extension but does require many firefox specific content security policy settings to be disabled in the firefox developer browser. All functionality matches chrome version. -BS